Fix/ tests in the midst of update
