+++ test/InMemoryTaskManagerTest.java
@@ -1,0 +1,13 @@
+import org.junit.jupiter.api.BeforeEach;
+import service.InMemoryTaskManager;
+import service.Managers;
+import service.TaskManager;
+
+public class InMemoryTaskManagerTest extends TaskManagerTest <InMemoryTaskManager> {
+//    @BeforeEach
+//    public void beforeEach() {
+//        taskManager = new InMemoryTaskManager();
+//        historyManager = Managers.getDefaultHistory();
+//    }
+}
+
