+++ test/TaskManagerTest.java
@@ -16,189 +16,66 @@ import java.util.List;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 
-public class TaskManagerTest {
+abstract class TaskManagerTest<T extends TaskManager> {
 
     public static TaskManager taskManager = Managers.getDefault();
     public static HistoryManager historyManager = Managers.getDefaultHistory();
+    private final Task task1 = new Task();
+    private final Epic epic1 = new Epic();
+    private final Subtask subtask1 = new Subtask();
+    private final Subtask subtask2 = new Subtask();
+
+
+
 
 
     @BeforeEach
     public void beforeEach() {
         taskManager = Managers.getDefault();
-        historyManager = Managers.getDefaultHistory();
-
-    }
-
-    // проверяем сохранение задач в отсутствующий/пустой файл
-    @Test
-    public void fileBackedTaskManagerShouldSaveTasksToNewFile() {
-        FileBackedTaskManager fm = new FileBackedTaskManager();
-        Task task1 = new Task();
-        task1.setTaskName("task1_name");
-        task1.setDescription("task1_description");
-        task1.setStatus(TaskStatus.NEW);
-        task1.setDuration(Duration.ofMinutes(30));
-        task1.setStartTime(LocalDateTime.of(2022,12,1,10,10,0));
-        task1 = fm.createNewTask(task1);
-        File saveFile = FileBackedTaskManager.getSaveFile();
-        Task middleTask = null;
-        try (BufferedReader fileReader = new BufferedReader(new FileReader(saveFile, StandardCharsets.UTF_8))) {
-            BufferedReader br = new BufferedReader(fileReader);
-
-            while (br.ready()) {
-                String line = br.readLine();
-                if (line.isEmpty()) {
-                    break;
-                }
-                if (!line.startsWith("id")) {
-                    middleTask = fm.fromString(line);
-                }
-            }
-        } catch (IOException e) {
-            throw new ManagerSaveException("Ошибка чтения файла");
-        }
-        assertEquals(fm.toString(task1), fm.toString(middleTask));
-        saveFile.delete();
-    }
-
-    // проверяем сохранение задач в отсутствующий/пустой файл
-    @Test
-    public void fileBackedTaskManagerShouldSaveMultipleTasks() {
-        FileBackedTaskManager fm = new FileBackedTaskManager();
-        Task task1 = new Task();
-        task1.setTaskName("task1_name");
-        task1.setDescription("task1_description");
-        task1.setStatus(TaskStatus.NEW);
-        task1.setDuration(Duration.ofMinutes(30));
-        task1.setStartTime(LocalDateTime.of(2022,12,1,10,10,0));
-        task1 = fm.createNewTask(task1);
-
-        Epic epic1 = new Epic();
-        epic1.setTaskName("epic1");
-        epic1.setDescription("epic1_description");
-        epic1 = fm.createNewEpic(epic1);
-
-        Subtask subtask1 = new Subtask();
-        subtask1.setTaskName("subtask1");
-        subtask1.setDescription("subtask1_description");
-        subtask1.setStatus(TaskStatus.NEW);
-        subtask1.setDuration(Duration.ofMinutes(15));
-        subtask1.setStartTime(LocalDateTime.of(2022,12,1,10,10,0));
-        subtask1.setEpicId(epic1.getId());
-        subtask1 = fm.createNewSubtask(subtask1);
-
-        Subtask subtask2 = new Subtask();
-        subtask2.setTaskName("subtask2");
-        subtask2.setDescription("subtask2_description");
-        subtask2.setStatus(TaskStatus.DONE);
-        subtask2.setDuration(Duration.ofMinutes(20));
-        subtask2.setStartTime(LocalDateTime.of(2022,12,1,10,25,0));
-        subtask2.setEpicId(epic1.getId());
-        subtask2 = fm.createNewSubtask(subtask2);
-        File saveFile = FileBackedTaskManager.getSaveFile();
-
-        List<String> taskStings = new ArrayList<>();
-        try (BufferedReader fileReader = new BufferedReader(new FileReader(saveFile, StandardCharsets.UTF_8))) {
-            BufferedReader br = new BufferedReader(fileReader);
-
-            while (br.ready()) {
-                String line = br.readLine();
-                if (line.isEmpty()) {
-                    break;
-                }
-                if (!line.startsWith("id")) {
-                    taskStings.add(line);
-                }
-            }
-        } catch (IOException e) {
-            throw new ManagerSaveException("Ошибка чтения файла");
-        }
-        assertEquals(taskStings.toString(), "[1,TASK,task1_name,NEW,task1_description,01.12.2022 - 10:10,30, " +
-                "2,EPIC,epic1,IN_PROGRESS,epic1_description,01.12.2022 - 10:10,35, 3,SUBTASK,subtask1," +
-                "NEW,subtask1_description,2,01.12.2022 - 10:10,15, 4,SUBTASK,subtask2,DONE,subtask2_description," +
-                "2,01.12.2022 - 10:25,20]");
-    }
-
-    @Test
-    public void fileBackedTaskManagerShouldLoadMultipleTasks() {
-        FileBackedTaskManager fm = new FileBackedTaskManager();
-        File testFile = FileBackedTaskManager.setLoadFile("tempTestFile.csv");
-
-        try (Writer fw = new FileWriter(testFile)) {
-            fw.write("id,type,name,status,description,epic,date-time,duration\n" +
-                    "1,TASK,task1_name,NEW,task1_description,01.12.2022 - 10:10,30\n" +
-                    "2,EPIC,epic1,IN_PROGRESS,epic1_description,01.12.2022 - 10:10,35\n" +
-                    "3,SUBTASK,subtask1,NEW,subtask1_description,2,01.12.2022 - 10:10,15\n" +
-                    "4,SUBTASK,subtask2,DONE,subtask2_description,2,01.12.2022 - 10:25,20\n");
-        } catch (IOException e) {
-            System.out.println("Ошибка записи в файл");
-        }
-        fm = FileBackedTaskManager.loadFromFile(testFile);
-        fm.save();
-        File saveFile = FileBackedTaskManager.getSaveFile();
-
-        List<String> taskStings = new ArrayList<>();
-        try (BufferedReader fileReader = new BufferedReader(new FileReader(saveFile, StandardCharsets.UTF_8))) {
-            BufferedReader br = new BufferedReader(fileReader);
-
-            while (br.ready()) {
-                String line = br.readLine();
-                if (line.isEmpty()) {
-                    break;
-                }
-                if (!line.startsWith("id")) {
-                    taskStings.add(line);
-                }
-            }
-        } catch (IOException e) {
-            throw new ManagerSaveException("Ошибка чтения файла");
-        }
-        assertEquals(taskStings.toString(), "[1,TASK,task1_name,NEW,task1_description,01.12.2022 - 10:10,30, " +
-                "2,EPIC,epic1,IN_PROGRESS,epic1_description,01.12.2022 - 10:10,35, 3,SUBTASK,subtask1," +
-                "NEW,subtask1_description,2,01.12.2022 - 10:10,15, 4,SUBTASK,subtask2,DONE,subtask2_description," +
-                "2,01.12.2022 - 10:25,20]");
-        testFile.delete();
-    }
-
-    // проверьте, что экземпляры класса Task равны друг другу, если равен их id
-    @Test
-    public void task1ShouldBeEqualToTask1InTaskManager() {
-        Task task1 = new Task();
+        //historyManager = Managers.getDefaultHistory();
         task1.setTaskName("task1_name");
         task1.setDescription("task1_description");
         task1.setStatus(TaskStatus.NEW);
         task1.setDuration(Duration.ofMinutes(20));
-        task1.setStartTime(LocalDateTime.of(2022,12,1,10,25,0));
-        task1 = taskManager.createNewTask(task1);
-        assertEquals(task1, taskManager.getTaskById(1));
-    }
+        task1.setStartTime(LocalDateTime.of(2022, 12, 1, 10, 25, 0));
 
-    // проверьте, что наследники класса Task равны друг другу, если равен их id
-    @Test
-    public void epic1ShouldBeEqualToEpic1InTaskManager() {
         Epic epic1 = new Epic();
         epic1.setTaskName("epic1");
         epic1.setDescription("epic1_description");
-        epic1 = taskManager.createNewEpic(epic1);
 
         Subtask subtask1 = new Subtask();
         subtask1.setTaskName("subtask1");
         subtask1.setDescription("subtask1_description");
         subtask1.setStatus(TaskStatus.NEW);
         subtask1.setDuration(Duration.ofMinutes(20));
-        subtask1.setStartTime(LocalDateTime.of(2022,12,1,10,25,0));
+        subtask1.setStartTime(LocalDateTime.of(2022, 12, 1, 10, 25, 0));
         subtask1.setEpicId(epic1.getId());
-        subtask1 = taskManager.createNewSubtask(subtask1);
 
         Subtask subtask2 = new Subtask();
         subtask2.setTaskName("subtask2");
         subtask2.setDescription("subtask2_description");
         subtask2.setStatus(TaskStatus.DONE);
         subtask2.setDuration(Duration.ofMinutes(0));
-        subtask2.setStartTime(LocalDateTime.of(2022,12,1,10,45,0));
+        subtask2.setStartTime(LocalDateTime.of(2022, 12, 1, 10, 45, 0));
         subtask2.setEpicId(epic1.getId());
-        subtask2 = taskManager.createNewSubtask(subtask2);
-        assertEquals(epic1, taskManager.getEpicById(1));
+
+
+    }
+
+    // проверьте, что экземпляры класса Task равны друг другу, если равен их id
+    @Test
+    public void task1ShouldBeEqualToTask1InTaskManager() {
+        taskManager.createNewTask(task1);
+        assertEquals(task1, taskManager.getTaskById(1));
+    }
+
+    // проверьте, что наследники класса Task равны друг другу, если равен их id
+    @Test
+    public void epic1ShouldBeEqualToEpic1InTaskManager() {
+        Epic testEpic = taskManager.createNewEpic(epic1);
+        taskManager.createNewSubtask(subtask1);
+        taskManager.createNewSubtask(subtask2);
+        assertEquals(testEpic, taskManager.getEpicById(1));
     }
 
     @Test
@@ -213,7 +90,7 @@ public class TaskManagerTest {
         subtask1.setDescription("subtask1_description");
         subtask1.setStatus(TaskStatus.NEW);
         subtask1.setDuration(Duration.ofMinutes(10));
-        subtask1.setStartTime(LocalDateTime.of(2022,12,1,10,35,0));
+        subtask1.setStartTime(LocalDateTime.of(2022, 12, 1, 10, 35, 0));
         subtask1.setEpicId(epic1.getId());
         subtask1 = taskManager.createNewSubtask(subtask1);
 
@@ -222,7 +99,7 @@ public class TaskManagerTest {
         subtask2.setDescription("subtask2_description");
         subtask2.setStatus(TaskStatus.DONE);
         subtask2.setDuration(Duration.ofMinutes(20));
-        subtask2.setStartTime(LocalDateTime.of(2022,12,1,10,45,0));
+        subtask2.setStartTime(LocalDateTime.of(2022, 12, 1, 10, 45, 0));
         subtask2.setEpicId(epic1.getId());
         subtask2 = taskManager.createNewSubtask(subtask2);
         assertEquals(subtask1, taskManager.getSubtaskById(2));
@@ -281,7 +158,7 @@ public class TaskManagerTest {
         task1.setDescription("task1_description");
         task1.setStatus(TaskStatus.NEW);
         task1.setDuration(Duration.ofMinutes(20));
-        task1.setStartTime(LocalDateTime.of(2022,12,1,10,45,0));
+        task1.setStartTime(LocalDateTime.of(2022, 12, 1, 10, 45, 0));
         task1 = taskManager.createNewTask(task1);
 
         Task task2 = new Task(); // id 2
@@ -289,7 +166,7 @@ public class TaskManagerTest {
         task2.setDescription("task2_description");
         task2.setStatus(TaskStatus.IN_PROGRESS);
         task2.setDuration(Duration.ofMinutes(20));
-        task2.setStartTime(LocalDateTime.of(2022,12,1,10,45,0));
+        task2.setStartTime(LocalDateTime.of(2022, 12, 1, 10, 45, 0));
         task2 = taskManager.createNewTask(task2);
 
         Task task3 = new Task(); // id 3
@@ -297,7 +174,7 @@ public class TaskManagerTest {
         task3.setDescription("task3_description");
         task3.setStatus(TaskStatus.IN_PROGRESS);
         task3.setDuration(Duration.ofMinutes(20));
-        task3.setStartTime(LocalDateTime.of(2022,12,1,10,45,0));
+        task3.setStartTime(LocalDateTime.of(2022, 12, 1, 10, 45, 0));
         task3 = taskManager.createNewTask(task3);
 
         Epic epic1 = new Epic(); // id 4
@@ -310,7 +187,7 @@ public class TaskManagerTest {
         subtask1.setDescription("subtask1_description");
         subtask1.setStatus(TaskStatus.NEW);
         subtask1.setDuration(Duration.ofMinutes(20));
-        subtask1.setStartTime(LocalDateTime.of(2022,12,1,10,45,0));
+        subtask1.setStartTime(LocalDateTime.of(2022, 12, 1, 10, 45, 0));
         subtask1.setEpicId(epic1.getId());
         subtask1 = taskManager.createNewSubtask(subtask1);
 
@@ -319,7 +196,7 @@ public class TaskManagerTest {
         subtask2.setDescription("subtask2_description");
         subtask2.setStatus(TaskStatus.DONE);
         subtask2.setDuration(Duration.ofMinutes(20));
-        subtask2.setStartTime(LocalDateTime.of(2022,12,1,11,5,0));
+        subtask2.setStartTime(LocalDateTime.of(2022, 12, 1, 11, 5, 0));
         subtask2.setEpicId(epic1.getId());
         subtask2 = taskManager.createNewSubtask(subtask2);
 
@@ -333,7 +210,7 @@ public class TaskManagerTest {
         subtask3.setDescription("subtask3_description");
         subtask3.setStatus(TaskStatus.DONE);
         subtask3.setDuration(Duration.ofMinutes(10));
-        subtask3.setStartTime(LocalDateTime.of(2022,12,1,11,25,0));
+        subtask3.setStartTime(LocalDateTime.of(2022, 12, 1, 11, 25, 0));
         subtask3.setEpicId(epic2.getId());
         subtask3 = taskManager.createNewSubtask(subtask3);
 
@@ -352,7 +229,7 @@ public class TaskManagerTest {
         subtask4.setDescription("subtask4_description");
         subtask4.setStatus(TaskStatus.NEW);
         subtask4.setDuration(Duration.ofMinutes(10));
-        subtask4.setStartTime(LocalDateTime.of(2022,12,1,11,35,0));
+        subtask4.setStartTime(LocalDateTime.of(2022, 12, 1, 11, 35, 0));
         subtask4.setEpicId(epic4.getId());
         subtask4 = taskManager.createNewSubtask(subtask4);
 
@@ -361,7 +238,7 @@ public class TaskManagerTest {
         subtask5.setDescription("subtask5_description");
         subtask5.setStatus(TaskStatus.NEW);
         subtask5.setDuration(Duration.ofMinutes(10));
-        subtask5.setStartTime(LocalDateTime.of(2022,12,1,11,45,0));
+        subtask5.setStartTime(LocalDateTime.of(2022, 12, 1, 11, 45, 0));
         subtask5.setEpicId(epic4.getId());
         subtask5 = taskManager.createNewSubtask(subtask5);
 
@@ -385,177 +262,12 @@ public class TaskManagerTest {
         task1.setDescription("task1_description");
         task1.setStatus(TaskStatus.NEW);
         task1.setDuration(Duration.ofMinutes(10));
-        task1.setStartTime(LocalDateTime.of(2022,12,1,11,45,0));
+        task1.setStartTime(LocalDateTime.of(2022, 12, 1, 11, 45, 0));
         task1 = taskManager.createNewTask(task1);
         assertEquals("task1_name", task1.getTaskName(1));
         assertEquals("task1_description", task1.getDescription());
         assertEquals(TaskStatus.NEW, task1.getStatus());
     }
-
-    // тест добавления в историю
-    @Test
-    void add() {
-        Task task1 = new Task(); // id 1
-        task1.setTaskName("task1_name");
-        task1.setDescription("task1_description");
-        task1.setStatus(TaskStatus.NEW);
-        task1.setDuration(Duration.ofMinutes(10));
-        task1.setStartTime(LocalDateTime.of(2022,12,1,11,45,0));
-        task1 = taskManager.createNewTask(task1);
-        taskManager.getTaskById(1);
-        final List<Task> history = historyManager.getHistory();
-        assertNotNull(history, "История не пустая.");
-        assertEquals(1, history.size(), "История не пустая.");
-    }
-
-    // убедитесь, что задачи, добавляемые в HistoryManager, сохраняют предыдущую версию задачи и её данных
-    @Test
-    public void historyShouldSaveSaveTaskAfterUpdate() {
-        Task task1 = new Task(); // id 1
-        task1.setTaskName("task1_name");
-        task1.setDescription("task1_description");
-        task1.setStatus(TaskStatus.NEW);
-        task1.setDuration(Duration.ofMinutes(10));
-        task1.setStartTime(LocalDateTime.of(2022,12,1,11,45,0));
-        task1 = taskManager.createNewTask(task1);
-        taskManager.getTaskById(1);
-        Task task1Updated = new Task();
-        task1Updated.setTaskName("task1_name_updated");
-        task1Updated.setDescription("task1_description_updated");
-        task1Updated.setStatus(TaskStatus.IN_PROGRESS);
-        task1Updated.setId(task1.getId());
-        task1Updated.setDuration(task1.getDuration());
-        task1Updated.setStartTime(task1.getStartTime());
-        taskManager.updateTask(task1Updated);
-        taskManager.getTaskById(1);
-        List<Task> history = historyManager.getHistory();
-
-        assertEquals("[Task{id='1 , taskName= task1_name_updated , description.length= 25, startTime= " +
-                        "01.12.2022 - 11:45, duration= 10, status=IN_PROGRESS }]",
-                history.toString());
-    }
-
-    @Test
-    public void historyShouldSaveUpdatedEpicsAndSubtasksFormAfterUpdate() {
-        Task task1 = new Task(); // id 1
-        task1.setTaskName("task1_name");
-        task1.setDescription("task1_description");
-        task1.setStatus(TaskStatus.NEW);
-        task1.setDuration(Duration.ofMinutes(10));
-        task1.setStartTime(LocalDateTime.of(2022,12,1,11,45,0));
-        task1 = taskManager.createNewTask(task1);
-
-        Task task2 = new Task(); // id 2
-        task2.setTaskName("task2_name");
-        task2.setDescription("task2_description");
-        task2.setStatus(TaskStatus.IN_PROGRESS);
-        task2.setDuration(Duration.ofMinutes(10));
-        task2.setStartTime(LocalDateTime.of(2022,12,1,11,45,0));
-        task2 = taskManager.createNewTask(task2);
-
-        Task task3 = new Task(); // id 3
-        task3.setTaskName("task3_name");
-        task3.setDescription("task3_description");
-        task3.setStatus(TaskStatus.IN_PROGRESS);
-        task3.setDuration(Duration.ofMinutes(10));
-        task3.setStartTime(LocalDateTime.of(2022,12,1,11,55,0));
-        task3 = taskManager.createNewTask(task3);
-
-        Epic epic1 = new Epic(); // id 4
-        epic1.setTaskName("epic1_name");
-        epic1.setDescription("epic1_description");
-        epic1 = taskManager.createNewEpic(epic1);
-
-        Subtask subtask1 = new Subtask(); // id 5
-        subtask1.setTaskName("subtask1_name");
-        subtask1.setDescription("subtask1_description");
-        subtask1.setStatus(TaskStatus.NEW);
-        subtask1.setDuration(Duration.ofMinutes(10));
-        subtask1.setStartTime(LocalDateTime.of(2022,12,1,12,5,0));
-        subtask1.setEpicId(epic1.getId());
-        subtask1 = taskManager.createNewSubtask(subtask1);
-
-        Subtask subtask2 = new Subtask(); // id 6
-        subtask2.setTaskName("subtask2_name");
-        subtask2.setDescription("subtask2_description");
-        subtask2.setStatus(TaskStatus.DONE);
-        subtask2.setDuration(Duration.ofMinutes(10));
-        subtask2.setStartTime(LocalDateTime.of(2022,12,1,12,15,0));
-        subtask2.setEpicId(epic1.getId());
-        subtask2 = taskManager.createNewSubtask(subtask2);
-
-        Epic epic2 = new Epic(); // id 7
-        epic2.setTaskName("epic2_name");
-        epic2.setDescription("epic2_description");
-        epic2 = taskManager.createNewEpic(epic2);
-
-        Subtask subtask3 = new Subtask(); // id 8
-        subtask3.setTaskName("subtask3_name");
-        subtask3.setDescription("subtask3_description");
-        subtask3.setStatus(TaskStatus.DONE);
-        subtask3.setDuration(Duration.ofMinutes(10));
-        subtask3.setStartTime(LocalDateTime.of(2022,12,1,12,25,0));
-        subtask3.setEpicId(epic2.getId());
-        subtask3 = taskManager.createNewSubtask(subtask3);
-
-        Epic epic3 = new Epic(); // id 9
-        epic3.setTaskName("epic3_name");
-        epic3.setDescription("epic3_description");
-        epic3 = taskManager.createNewEpic(epic3);
-
-        Epic epic4 = new Epic(); // id 10
-        epic4.setTaskName("epic4_name");
-        epic4.setDescription("epic4_description");
-        epic4 = taskManager.createNewEpic(epic4);
-
-        Subtask subtask4 = new Subtask(); // id 11
-        subtask4.setTaskName("subtask4_name");
-        subtask4.setDescription("subtask4_description");
-        subtask4.setStatus(TaskStatus.NEW);
-        subtask4.setDuration(Duration.ofMinutes(10));
-        subtask4.setStartTime(LocalDateTime.of(2022,12,1,12,35,0));
-        subtask4.setEpicId(epic4.getId());
-        subtask4 = taskManager.createNewSubtask(subtask4);
-
-        Subtask subtask5 = new Subtask(); // id 12
-        subtask5.setTaskName("subtask5_name");
-        subtask5.setDescription("subtask5_description");
-        subtask5.setStatus(TaskStatus.NEW);
-        subtask5.setDuration(Duration.ofMinutes(10));
-        subtask5.setStartTime(LocalDateTime.of(2022,12,1,12,45,0));
-        subtask5.setEpicId(epic4.getId());
-        subtask5 = taskManager.createNewSubtask(subtask5);
-
-        String epic2String = "[Epic{id='7 , taskName= epic2_updated_name , description.length=25, startTime= " +
-                "01.12.2022 - 12:25, duration= 10, status=DONE subtasksIds= [8]}]";
-        taskManager.getEpicById(7);
-        Epic epic2Updated = new Epic();
-        epic2Updated.setTaskName("epic2_updated_name");
-        epic2Updated.setDescription("epic2_updated_description");
-        epic2Updated.setId(epic2.getId());
-        epic2Updated.setStatus(TaskStatus.NEW);
-        epic2Updated.setDuration(epic2.getDuration());
-        epic2Updated.setStartTime(epic2.getStartTime());
-
-        taskManager.updateEpic(epic2Updated);
-        taskManager.getEpicById(7);
-        assertEquals(epic2String, historyManager.getHistory().toString());
-
-        taskManager.getSubtaskById(12);
-        Subtask subtask5Updated = new Subtask();
-        subtask5Updated.setTaskName("subtask5_updated_name");
-        subtask5Updated.setDescription("subtask5_updated_description");
-        subtask5Updated.setStatus(TaskStatus.IN_PROGRESS);
-        subtask5Updated.setDuration(subtask5.getDuration());
-        subtask5Updated.setStartTime(subtask5.getStartTime());
-        subtask5Updated.setId(subtask5.getId());
-        subtask5Updated.setEpicId(epic4.getId());
-
-        taskManager.updateSubTask(subtask5Updated);
-        String subtask5String = "[Epic{id='7 , taskName= epic2_updated_name , description.length=25, startTime= " +
-                "01.12.2022 - 12:25, duration= 10, status=DONE subtasksIds= [8]}, Subtask{id='12 , taskName= " +
-                "subtask5_name , description.length=20, startTime= 01.12.2022 - 12:45, duration= 10, status=NEW " +
-                "epicId = 10}]";
-        assertEquals(subtask5String, historyManager.getHistory().toString());
-    }
 }
+
+
